static_extension_info_t: Use uint8_t again were possible
The previous commit changed these fields into uint16_t since xproto tries to
assign value 256 here. However, this isn't actually necessary, so this commit
reverts that and instead makes the code generator emit a 0 here, since this
value isn't used anyway.

+++ src/errors.h
@@ -30,8 +30,8 @@
 
 struct static_extension_info_t {
 	uint16_t num_minor;
-	uint16_t num_events;
-	uint16_t num_errors;
+	uint8_t num_events;
+	uint8_t num_errors;
 	const char *strings_minor;
 	const char *strings_events;
 	const char *strings_errors;
